moving three components to own files under components directory, audio can optionally play slower
